extract sanitation_transforms.rb
